requirements.txt
python-dotenv
langchain
langchain_community
langchain_openai
numexpr
openai
flask
colorlog
pillow
httpx==0.27.2